+++ crates/icy_board_engine/src/icy_board/pcb/mod.rs
@@ -14,6 +14,7 @@ use self::{
 
 use super::{IcyBoardError, icb_text::IcbTextFile, lookup_case_insensitive};
 
+pub mod origins;
 pub mod pcbconferences;
 pub mod pcboard_data;
 pub mod user_inf;